helper updated en helper functies in RouteMenu

+++ Project/Logic/Helper.cs
@@ -15,7 +15,7 @@ public static class Helper
 
     public static bool IsDouble(object parameter)
     {
-        return parameter.GetType() == typeof(int);
+        return parameter.GetType() == typeof(double);
     }
